Simplify waitForRulesToFinishExecution helper function
