+++ x-pack/plugins/security_solution/cypress/tasks/api_calls/rules.ts
@@ -7,9 +7,13 @@
 
 import moment from 'moment';
 import { rootRequest } from '../common';
-import { DETECTION_ENGINE_RULES_URL } from '../../../common/constants';
+import {
+  DETECTION_ENGINE_RULES_URL,
+  DETECTION_ENGINE_RULES_URL_FIND,
+} from '../../../common/constants';
 import type { RuleCreateProps, RuleResponse } from '../../../common/detection_engine/rule_schema';
 import { internalAlertingSnoozeRule } from '../../urls/routes';
+import type { FetchRulesResponse } from '../../../public/detection_engine/rule_management/logic/types';
 
 export const createRule = (
   rule: RuleCreateProps
@@ -72,3 +76,20 @@ export const importRule = (ndjsonPath: string) => {
         .should('be.equal', 200);
     });
 };
+
+export const waitForRulesToFinishExecution = (ruleIds: string[]) =>
+  cy.waitUntil(
+    () =>
+      rootRequest<FetchRulesResponse>({
+        method: 'GET',
+        url: DETECTION_ENGINE_RULES_URL_FIND,
+      }).then((response) => {
+        const areAllRulesFinished = ruleIds.every((ruleId) =>
+          response.body.data.some(
+            (rule) => rule.rule_id === ruleId && typeof rule.execution_summary !== 'undefined'
+          )
+        );
+        return areAllRulesFinished;
+      }),
+    { interval: 500, timeout: 12000 }
+  );
